+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+v0.6.0 (2017-03-10)
+-------------------
+* Added new `drop_dose` method to Dataset
+* don't attempt running models when there are too few dose-groups
+* removed doses-dropped parameter from dataset init
+* Add example notebook running actual data
+
 v0.5.3 (2017-03-02)
 -------------------